# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Grantsy Entitlements API",
    "description": "Microservice for managing SaaS feature entitlements",
    "version": "1.0.0"
  },
  "paths": {
    "/v1/check": {
      "get": {
        "tags": [
          "Entitlements"
        ],
        "summary": "Check feature access",
        "description": "Check if a user has access to a specific feature based on their subscription plan. Use ?expand=feature\u0026expand=plan\u0026expand=plan.features to include additional details.",
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "description": "User ID to check access for",
            "schema": {
              "type": "string",
              "description": "User ID to check access for"
            }
          },
          {
            "name": "feature",
            "in": "query",
            "description": "Feature ID to check access for",
            "schema": {
              "type": "string",
              "description": "Feature ID to check access for"
            }
          },
          {
            "name": "expand",
            "in": "query",
            "description": "Fields to expand (use ?expand=feature\u0026expand=plan\u0026expand=plan.features)",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/CheckExpand"
              },
              "description": "Fields to expand (use ?expand=feature\u0026expand=plan\u0026expand=plan.features)"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feature access check result",
            "content": {
              "application/json": {
                "schema": {
                  "title": "CheckResponse",
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/CheckResponse"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/features": {
      "get": {
        "tags": [
          "Features"
        ],
        "summary": "List all features",
        "description": "Get all available feature definitions",
        "responses": {
          "200": {
            "description": "All available features",
            "content": {
              "application/json": {
                "schema": {
                  "title": "FeaturesResponse",
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/FeaturesResponse"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/features/{feature_id}": {
      "get": {
        "tags": [
          "Features"
        ],
        "summary": "Get feature by ID",
        "description": "Get details of a specific feature by its identifier",
        "parameters": [
          {
            "name": "feature_id",
            "in": "path",
            "description": "Feature ID to look up",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Feature ID to look up"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Feature details",
            "content": {
              "application/json": {
                "schema": {
                  "title": "FeatureResponse",
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/FeatureResponse"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/plans": {
      "get": {
        "tags": [
          "Plans"
        ],
        "summary": "List all plans",
        "description": "Get all available subscription plans with their pricing variants. Use ?expand=features to include features.",
        "parameters": [
          {
            "name": "expand",
            "in": "query",
            "description": "Fields to expand (use ?expand=features)",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/PlansExpand"
              },
              "description": "Fields to expand (use ?expand=features)"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of available plans",
            "content": {
              "application/json": {
                "schema": {
                  "title": "PlansResponse",
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PlansResponse"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/plans/{plan_id}": {
      "get": {
        "tags": [
          "Plans"
        ],
        "summary": "Get plan by ID",
        "description": "Get details of a specific plan by its identifier. Use ?expand=features to include feature details.",
        "parameters": [
          {
            "name": "expand",
            "in": "query",
            "description": "Fields to expand (use ?expand=features)",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/PlanExpand"
              },
              "description": "Fields to expand (use ?expand=features)"
            }
          },
          {
            "name": "plan_id",
            "in": "path",
            "description": "Plan ID to look up",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Plan ID to look up"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan details",
            "content": {
              "application/json": {
                "schema": {
                  "title": "PlanResponse",
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PlanResponse"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    },
    "/v1/users/{user_id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user state",
        "description": "Get the current state for a user. Always returns plan_id. Use ?expand=plan,features,subscription to include additional details.",
        "parameters": [
          {
            "name": "expand",
            "in": "query",
            "description": "Fields to expand (use ?expand=plan\u0026expand=features\u0026expand=subscription)",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/UserExpand"
              },
              "description": "Fields to expand (use ?expand=plan\u0026expand=features\u0026expand=subscription)"
            }
          },
          {
            "name": "user_id",
            "in": "path",
            "description": "User ID to look up",
            "required": true,
            "schema": {
              "type": "string",
              "description": "User ID to look up"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User state",
            "content": {
              "application/json": {
                "schema": {
                  "title": "UserResponse",
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/UserResponse"
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - missing or invalid API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "ApiKeyAuth": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
      "CheckExpand": {
        "enum": [
          "feature",
          "plan",
          "plan.features"
        ],
        "type": "string"
      },
      "CheckResponse": {
        "type": "object",
        "properties": {
          "allowed": {
            "type": "boolean",
            "description": "Whether the user has access to this feature"
          },
          "feature": {
            "$ref": "#/components/schemas/Feature"
          },
          "plan": {
            "$ref": "#/components/schemas/Plan"
          },
          "reason": {
            "enum": [
              "no_subscription",
              "default_plan",
              "feature_in_plan",
              "insufficient_plan"
            ],
            "type": "string",
            "description": "Reason for the access decision"
          },
          "user_id": {
            "type": "string",
            "description": "The user ID"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ProblemDetails"
          }
        }
      },
      "Feature": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Feature description"
          },
          "id": {
            "type": "string",
            "description": "Feature identifier"
          },
          "name": {
            "type": "string",
            "description": "Feature display name"
          }
        }
      },
      "FeatureResponse": {
        "type": "object",
        "properties": {
          "feature": {
            "$ref": "#/components/schemas/Feature"
          }
        }
      },
      "FeaturesResponse": {
        "type": "object",
        "properties": {
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Feature"
            },
            "description": "All available features",
            "nullable": true
          }
        }
      },
      "FieldError": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "LemonSqueezySubscription": {
        "type": "object",
        "properties": {
          "billing_anchor": {
            "type": "integer",
            "description": "Day of month for billing"
          },
          "cancelled": {
            "type": "boolean",
            "description": "Whether the subscription has been cancelled"
          },
          "card_brand": {
            "type": "string",
            "description": "Payment card brand"
          },
          "card_last_four": {
            "type": "string",
            "description": "Last four digits of payment card"
          },
          "created_at": {
            "type": "integer",
            "description": "Unix timestamp when subscription was created",
            "format": "int64"
          },
          "customer_id": {
            "type": "integer",
            "description": "LemonSqueezy customer ID"
          },
          "ends_at": {
            "type": "integer",
            "description": "Unix timestamp when subscription ends",
            "nullable": true
          },
          "id": {
            "type": "integer",
            "description": "LemonSqueezy subscription ID"
          },
          "order_id": {
            "type": "integer",
            "description": "LemonSqueezy order ID"
          },
          "product_id": {
            "type": "integer",
            "description": "LemonSqueezy product ID"
          },
          "product_name": {
            "type": "string",
            "description": "Product display name"
          },
          "renews_at": {
            "type": "integer",
            "description": "Unix timestamp when subscription renews",
            "format": "int64"
          },
          "status": {
            "type": "string",
            "description": "Subscription status"
          },
          "status_formatted": {
            "type": "string",
            "description": "Human-readable subscription status"
          },
          "subscription_item_id": {
            "type": "integer",
            "description": "LemonSqueezy subscription item ID"
          },
          "trial_ends_at": {
            "type": "integer",
            "description": "Unix timestamp when trial ends",
            "nullable": true
          },
          "updated_at": {
            "type": "integer",
            "description": "Unix timestamp when subscription was last updated",
            "format": "int64"
          },
          "variant_id": {
            "type": "integer",
            "description": "LemonSqueezy variant ID"
          },
          "variant_name": {
            "type": "string",
            "description": "Variant display name"
          }
        }
      },
      "Meta": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        }
      },
      "Plan": {
        "type": "object",
        "properties": {
          "description": {
            "type": "string",
            "description": "Plan description"
          },
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Feature"
            },
            "description": "Features included in this plan"
          },
          "id": {
            "type": "string",
            "description": "Plan identifier"
          },
          "name": {
            "type": "string",
            "description": "Plan display name"
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Variant"
            },
            "description": "Pricing variants for this plan"
          }
        }
      },
      "PlanExpand": {
        "enum": [
          "features"
        ],
        "type": "string"
      },
      "PlanResponse": {
        "type": "object",
        "properties": {
          "plan": {
            "$ref": "#/components/schemas/Plan"
          }
        }
      },
      "PlansExpand": {
        "enum": [
          "features"
        ],
        "type": "string"
      },
      "PlansResponse": {
        "type": "object",
        "properties": {
          "plans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Plan"
            },
            "description": "List of available plans",
            "nullable": true
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "detail": {
            "type": "string"
          },
          "fields": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            }
          },
          "request_id": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          },
          "title": {
            "type": "string"
          },
          "type": {
            "enum": [
              "https://grantsy.example/errors/validation-failed",
              "https://grantsy.example/errors/bad-request",
              "https://grantsy.example/errors/not-found",
              "https://grantsy.example/errors/unauthorized",
              "https://grantsy.example/errors/internal-error"
            ],
            "type": "string"
          }
        }
      },
      "ProviderSubscription": {
        "type": "object",
        "oneOf": [
          {
            "$ref": "#/components/schemas/LemonSqueezySubscription"
          }
        ]
      },
      "RawSubscription": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ProviderSubscription"
          },
          "provider": {
            "enum": [
              "lemonsqueezy"
            ],
            "type": "string",
            "description": "Provider identifier"
          }
        }
      },
      "UserExpand": {
        "enum": [
          "plan",
          "features",
          "subscription"
        ],
        "type": "string"
      },
      "UserResponse": {
        "type": "object",
        "properties": {
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Feature"
            },
            "description": "Features available to the user (requires expand=features)"
          },
          "plan": {
            "$ref": "#/components/schemas/Plan"
          },
          "plan_id": {
            "type": "string",
            "description": "The user's current plan ID"
          },
          "subscription": {
            "$ref": "#/components/schemas/UserSubscription"
          },
          "user_id": {
            "type": "string",
            "description": "The user ID"
          }
        }
      },
      "UserSubscription": {
        "type": "object",
        "properties": {
          "cancelled": {
            "type": "boolean",
            "description": "Whether the subscription has been cancelled"
          },
          "ends_at": {
            "type": "integer",
            "description": "Unix timestamp when subscription ends (if cancelled)",
            "nullable": true
          },
          "raw": {
            "$ref": "#/components/schemas/RawSubscription"
          },
          "renews_at": {
            "type": "integer",
            "description": "Unix timestamp when subscription renews",
            "nullable": true
          },
          "status": {
            "type": "string",
            "description": "Subscription status (active, on_trial, paused, past_due, cancelled, expired)"
          },
          "trial_ends_at": {
            "type": "integer",
            "description": "Unix timestamp when trial ends (if on trial)",
            "nullable": true
          }
        }
      },
      "Variant": {
        "type": "object",
        "properties": {
          "has_free_trial": {
            "type": "boolean",
            "description": "Whether this variant has a free trial"
          },
          "id": {
            "type": "integer",
            "description": "Variant identifier"
          },
          "interval": {
            "type": "string",
            "description": "Billing interval (month, year, etc.)"
          },
          "interval_count": {
            "type": "integer",
            "description": "Number of intervals between billings"
          },
          "name": {
            "type": "string",
            "description": "Variant display name"
          },
          "price": {
            "description": "Price in cents"
          },
          "sort": {
            "type": "integer",
            "description": "Display order"
          },
          "trial_interval": {
            "type": "string",
            "description": "Trial billing interval"
          },
          "trial_interval_count": {
            "type": "integer",
            "description": "Trial duration in intervals"
          }
        }
      }
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "name": "X-Api-Key",
        "in": "header",
        "description": "API key for authentication"
      }
    }
  }
}